requirements.txt
tensorflow
matplotlib
IPython
keras
fastapi
uvicorn
numpy
Pillow